clean up trigger event handlers
